changed file layout and changed test for instance of dictionary
