--- Braille.txt
@@ -1,3 +1,0 @@
-0.0.0.0.0....00.0.0.00
-00.00.0..0..00.0000..0
-....0.0.0....00.0.0...
