+++ lib/dictionary.rb
@@ -31,15 +31,15 @@ class Dictionary
                    "x" => ["00","..","00"],
                    "y" => ["00",".0","00"],
                    "z" => ["0.",".0","00"],
-                    " " => ["..","..",".."]
+                   " " => ["..","..",".."]
                  }
   end
 
   def number_of_keys
-   @dictionary.length
+   @library.length
   end
 
   def class_check
-    @dictionary.class
+    @library.class
   end
 end
