+++ lib/message.txt
@@ -1,1 +1,1 @@
-hello
+j
